+++ packages/espressocash_app/lib/features/intercom/services/intercom_service.dart
@@ -36,6 +36,10 @@ class IntercomService implements Disposable {
   void updateCountry(String? countryCode) => Intercom.instance
       .updateUser(customAttributes: {'countryCode': countryCode});
 
+  void updateStellarAddress(String address) => Intercom.instance.updateUser(
+        customAttributes: {'stellarAddress': address},
+      );
+
   @override
   Future<void> onDispose() => Intercom.instance.logout();
 }