+++ packages/solana/lib/src/rpc/extension.dart
@@ -9,9 +9,9 @@ extension RpcClientExt on RpcClient {
     FutureOr<void> Function(Signature signature)? onSigned,
     Commitment commitment = Commitment.finalized,
   }) async {
-    final recentBlockhash =
-        await getRecentBlockhash(commitment: commitment).value;
-    final signedTx = await signTransaction(recentBlockhash, message, signers);
+    final bh = await getLatestBlockhash(commitment: commitment).value;
+
+    final signedTx = await signTransaction(bh, message, signers);
 
     if (onSigned != null) {
       await onSigned(signedTx.signatures.first);
@@ -26,13 +26,13 @@ extension RpcClientExt on RpcClient {
   Future<SignedTx> signMessage(
     Message message,
     List<Ed25519HDKeyPair> signers, {
-    RecentBlockhash? blockhash,
+    LatestBlockhash? blockhash,
     Commitment commitment = Commitment.finalized,
   }) async {
-    final recentBlockhash =
-        blockhash ?? await getRecentBlockhash(commitment: commitment).value;
+    final latestBlockhash =
+        blockhash ?? await getLatestBlockhash(commitment: commitment).value;
 
-    return signTransaction(recentBlockhash, message, signers);
+    return signTransaction(latestBlockhash, message, signers);
   }
 
   /// Get the [limit] most recent transactions for the [address] account